Some small fixes

+++ src/main/java/at/favre/tools/dice/RndTool.java
@@ -256,7 +256,7 @@ public final class RndTool {
 
     @NotNull
     private static String getSummary(long durationMs, long durationRndGen, long byteGen) {
-        double bandwidth = durationRndGen == 0 || byteGen == 0 ? 0 : Math.round(byteGen / durationRndGen / 10.24) / 100.0;
+        double bandwidth = durationRndGen == 0 || byteGen == 0 ? 0 : Math.round((double) byteGen / (double) durationRndGen / 10.24) / 100.0;
         return System.lineSeparator() + System.lineSeparator() + "[" + getFriendlyFormattedDate() + "][" + MiscUtil.jarVersion() + "] " + byteGen + " bytes generated in " + durationMs + " ms." + (bandwidth > 0 ? " (" + bandwidth + " MB/s)" : "");
     }
